27-oct-2005 NvE Memberfunction GetX to access individual vector components also introduced                 for Ali4Vector.                 Also memberfunction SortJets introduced in AliVertex.                 Misleading comment modified in AliDevice.cxx. 29-oct-2005 NvE Forgotten angular unit setting fixed w.r.t. begin point etc... in AliTrack::ListAll                 ad begin point etc... info also added in AliTrack::List.                 Memberfunction SortTracks introduced in AliJet and comment corrected in AliVertex.h. 31-oct-2005 NvE Slight correction in IceCalibrate.cxx and IceCleanhits.cxx to satisfy                 the gcc compiler (thanks Dipo). 31-oct-2005 NvE Some unused variables removed in IceCal2Root and IceF2k to prevent warnings                 from the gcc compiler.

+++ RALICE/icepack/IceCleanHits.cxx
@@ -152,7 +152,7 @@ void IceCleanHits::Amanda()
  // It seems that in 2005 the trigger time was changed within the year
  // from 24170 ns to 12138 ns. The latter however shows a 2-bump structure,
  // so currently the 24170 ns will be used for the 2005 data.
- Int_t year=fEvt->GetJE();
+ Int_t year=(int)fEvt->GetJE();
  Float_t ttrig=23958;
  if (year==2003) ttrig=23994;
  if (year==2004) ttrig=24059.5;
